Componentized the folder puzzle
